Update init logic
Change the order of operations so that first the
cells are initialized to 0 or 128, then copy the
padding cells, then run the initializeNeighbor
Count function. Finally copy the padding
cells once more since the values they get from
initialzeNeighborCount are not correct even though
they are required to be considered for the
initializeNeighborCount function to work.